unified id generation experience
