+++ sgl-router/src/routers/openai/utils.rs
@@ -97,25 +97,10 @@ impl OutputIndexMapper {
 // ID Generation
 // ============================================================================
 
-/// Generate a unique ID with a prefix (similar to OpenAI format)
-///
-/// Generates IDs in the format: `{prefix}_{hex_string}` where hex_string is
-/// 50 hex characters (25 random bytes).
-///
-/// Examples:
-/// - `generate_id("msg")` → "msg_1234567890ab..."
-pub fn generate_id(prefix: &str) -> String {
-    use rand::RngCore;
-    let mut rng = rand::rng();
-    // Generate exactly 50 hex characters (25 bytes) for the part after the underscore
-    let mut bytes = [0u8; 25];
-    rng.fill_bytes(&mut bytes);
-    let hex_string: String = bytes.iter().map(|b| format!("{:02x}", b)).collect();
-    format!("{}_{}", prefix, hex_string)
-}
+// Re-export generate_id from protocols::common for backward compatibility
+pub use crate::protocols::common::generate_id;
 
 // ============================================================================
 // Re-export FunctionCallInProgress from mcp module
 // ============================================================================
-
 pub(crate) use super::mcp::FunctionCallInProgress;
